backend/requirements.txt
django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djoser==2.2.0
drf-spectacular==0.26.4
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
